# data_apis/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorBody": {
				"description": "Standard error response body returned by the error handler middleware.",
				"properties": {
					"error": {
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"message",
							"code"
						],
						"type": "object"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"TimeObjectResponse": {
				"properties": {
					"timestamp": {
						"type": "string"
					},
					"timezone": {
						"type": "string"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"duration_unit": {
						"type": "string"
					}
				},
				"required": [
					"timestamp",
					"timezone"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventRecordResponse_Record_string.unknown__": {
				"properties": {
					"event_id": {
						"type": "string"
					},
					"time_object": {
						"$ref": "#/components/schemas/TimeObjectResponse"
					},
					"event_type": {
						"type": "string"
					},
					"attribute": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"event_id",
					"time_object",
					"event_type",
					"attribute"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventDatasetResponse": {
				"properties": {
					"data_source": {
						"type": "string"
					},
					"dataset_type": {
						"type": "string"
					},
					"dataset_id": {
						"type": "string"
					},
					"time_object": {
						"$ref": "#/components/schemas/TimeObjectResponse"
					},
					"events": {
						"items": {
							"$ref": "#/components/schemas/EventRecordResponse_Record_string.unknown__"
						},
						"type": "array"
					}
				},
				"required": [
					"data_source",
					"dataset_type",
					"dataset_id",
					"time_object",
					"events"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventTypesResponse": {
				"properties": {
					"event_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"event_types"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventStatGroup": {
				"properties": {
					"key": {
						"type": "string"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"key",
					"count"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventStatsResponse": {
				"properties": {
					"total_events": {
						"type": "number",
						"format": "double"
					},
					"groups": {
						"items": {
							"$ref": "#/components/schemas/EventStatGroup"
						},
						"type": "array"
					}
				},
				"required": [
					"total_events",
					"groups"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EventRecordResponse": {
				"properties": {
					"event_id": {
						"type": "string"
					},
					"time_object": {
						"$ref": "#/components/schemas/TimeObjectResponse"
					},
					"event_type": {
						"type": "string"
					},
					"attribute": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"event_id",
					"time_object",
					"event_type",
					"attribute"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "esg-data-service",
		"version": "0.1.0",
		"description": "ESG CSV Ingestion, Retrieval & Preprocessing — single-container multi-domain API + SQS Worker",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/events": {
			"get": {
				"operationId": "GetEvents",
				"responses": {
					"200": {
						"description": "List of ESG metric events",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EventDatasetResponse"
								}
							}
						}
					},
					"501": {
						"description": "Not yet implemented",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorBody"
								}
							}
						}
					}
				},
				"description": "Query normalized ESG metric events from the data lake.\nSupports filtering by company, metric name, ESG pillar, and year range.\nResults are paginated via limit/offset.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "company_name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "permid",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "metric_name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Environmental | Social | Governance",
						"in": "query",
						"name": "pillar",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "year_from",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "year_to",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/events/types": {
			"get": {
				"operationId": "GetEventTypes",
				"responses": {
					"200": {
						"description": "Array of distinct event type strings",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EventTypesResponse"
								}
							}
						}
					},
					"501": {
						"description": "Not yet implemented",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorBody"
								}
							}
						}
					}
				},
				"description": "Returns the list of distinct event_type values present in the ingested dataset.\nUseful for populating filter dropdowns in the reporting frontend.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/events/stats": {
			"get": {
				"operationId": "GetEventStats",
				"responses": {
					"200": {
						"description": "Aggregated statistics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EventStatsResponse"
								}
							}
						}
					},
					"501": {
						"description": "Not yet implemented",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorBody"
								}
							}
						}
					}
				},
				"description": "Returns aggregate statistics over the ESG event dataset.\nSupports grouping by pillar, company, year, or industry.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"description": "Dimension to group by: pillar | company_name | metric_year | industry",
						"in": "query",
						"name": "group_by",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/events/{eventId}": {
			"get": {
				"operationId": "GetEventById",
				"responses": {
					"200": {
						"description": "A single ESG metric event record",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EventRecordResponse"
								}
							}
						}
					},
					"404": {
						"description": "Event not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorBody"
								}
							}
						}
					},
					"501": {
						"description": "Not yet implemented",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorBody"
								}
							}
						}
					}
				},
				"description": "Retrieves a single ESG EventRecord by its unique identifier.\nReturns the time_object, event_type, and full ESG metric attribute payload.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	]
}